Revert "Revert "{EQEmu PR} INT64""
This reverts commit b92f9582d5b2bd5d20c14c9519e549fabc209281.

+++ common/string_util.h
@@ -205,7 +205,7 @@ char *CleanMobName(const char *in, char *out);
 char *RemoveApostrophes(const char *s);
 void replace_all(std::string& in, std::string old, std::string repl);
 char* strn0cpy(char* dest, const char* source, uint32 size);
-const char *ConvertArray(int input, char *returnchar);
+const char *ConvertArray(int64 input, char *returnchar);
 const char *ConvertArrayF(float input, char *returnchar);
 const char *MakeLowerString(const char *source);
 uint32 hextoi(const char* num);
